requirements.txt
flask[async]
pact-python
yarl
pytest
python-dotenv
requests